New django portal app with social login
The new app is intended to integrate all the different applications in the
portal. It has implemented OAuth2 login/logut/settings/signup and tested
with GitHub. Integration with FIWARE missing.

The old COEGGS portal has been refactored into a folder and will be used
as a reference for what will be done in that project.